moving logic back to component
