+++ src/pages/order-history/order-history.js
@@ -1,5 +1,6 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { useTranslation } from 'react-i18next';
+import { useQuery } from '@apollo/client';
 import OrderHistoryOrder from '../../containers/orders/order-history/order-history-order';
 import EmptyOrderHistory from '../../containers/orders/order-history/empty-order-history';
 import OrderHistoryPagination from '../../containers/orders/order-history/order-history-pagination/index';
@@ -7,15 +8,36 @@ import { useStyles } from './order-history.styles';
 import { limitHistoryOrders } from '../../const/user-order-history';
 import errorOrLoadingHandler from '../../utils/errorOrLoadingHandler';
 import { getUserOrdersQuery } from './operations/order-history.queries';
-import usePaginationForOrders from './hooks/use-pagination-for-orders';
 
 const OrderHistory = () => {
-  const { loadingOrders, errorOrders, orders, currentPage, changeHandler, quantityPages } =
-    usePaginationForOrders(limitHistoryOrders, getUserOrdersQuery);
   const { t } = useTranslation();
   const styles = useStyles();
+  const [currentPage, setCurrentPage] = useState(1);
+
+  const {
+    loading: loadingOrders,
+    error: errorOrders,
+    data
+  } = useQuery(getUserOrdersQuery, {
+    variables: {
+      pagination: {
+        limit: limitHistoryOrders,
+        skip: (currentPage - 1) * limitHistoryOrders
+      }
+    },
+    fetchPolicy: 'network-only',
+    nextFetchPolicy: 'cache-first'
+  });
 
   if (loadingOrders || errorOrders) return errorOrLoadingHandler(errorOrders, loadingOrders);
+
+  const orders = data.getUserOrders.userOrders;
+  const { ordersCount } = data.getUserOrders;
+  const quantityPages = Math.ceil(ordersCount / limitHistoryOrders);
+
+  const changeHandler = (value) => {
+    setCurrentPage(value);
+  };
 
   return (
     <div className={styles.root}>
