+++ src/pages/order-history/tests/order-history.spec.js
@@ -1,15 +1,15 @@
 import React from 'react';
-import { useQuery } from '@apollo/client';
 import OrderHistory from '../order-history';
 
 jest.mock('../../order-history/order-history.styles', () => ({ useStyles: () => ({}) }));
-jest.mock('@apollo/client');
-
-useQuery.mockImplementation(() => ({}));
+jest.mock('@apollo/client', () => ({
+  ...jest.requireActual('@apollo/client'),
+  useQuery: () => ({ loading: false, error: null, data: { getUserOrders: {} } })
+}));
 
 describe('OrderHistory component', () => {
   const wrapper = shallow(<OrderHistory />);
-  it('should render', () => {
+  it('should render component', () => {
     expect(wrapper).toBeDefined();
   });
 });
